Added Version Check for VLFs, Changed to table variable
When this script fails, it usually fails on the VLF check and leaves a
temp table behind, which causes mass fumbling.

Added a version check in case the script is run all at once (or the
wrong check is run by accident).

Also switched the VLF check to use table variables for easier
re-runnability in case the wrong script is highlighted without the
version check (which could still happen).