Fixes @IgnoreQueryHashes bug
This wasn't working because it could only filter results in the first
run through dynamic SQL. Also, the two tables that we're actually
retrieving column from needed to be aliased in the select list to not
pull the query_hash column from the temp table to ignore, which would
give you two columns named query_hash.

In the second run through, there is no column named query_hash to filter
on.

In the third pass, Jeremiah explicitly referenced NULL as QueryHash, so
it wasn't going to filter on anything anyway.

This runs without error now when you ignore query hashes.

But this begs the question, why did he set up ignores for this, but not
sql_handle? Heh.