Get stored procedure cost
This is a little bit bobo. We update the QueryCost column twice -- once
correctly gets statement level costs, the other correctly gets stored
procedure level costs, and can differentiate between runs of the same
stored procedure.

Can't figure out a way to do it all in one shot that works correctly.
Either you miss the stored proc differentiation, or the statement level
costs are incorrect. It's a little inefficient, but it's still fast.