Sonic Debugger
Where to begin with this one? It turns out that setting the query hash
in the aggregated update was messing stuff up downstream. I had to break
out an old version of BC to catch it. When I added in my new code bit by
bit, it was easier to see what I did wrong. At one point the old BC with
my code was updating costs perfectly. QueryHash was the key. After I
took that out, doing the aggregations for statement level checks became
unnecessary. I'd rather have the stored procedure costs than the less
interesting information that we got there. There was also an important
update way down the bottom which was correctly setting statement level
costs. Now this reliably gets both, and keeps all the other fixes and
changes.

Really sorry for all the pull requests and pulling pull requests. This
thing is a bear to make big changes to.