Change Remote Query Cost warning
Reasoning for this: Want to warn for remote queries if the remote query
operator is 5% or more of the plan cost, regardless of plan cost. Remote
queries are horrible. The reason I'm setting it so low is because I've
run across instances where there are multiple remote servers being hit
in a plan with sporadic costs.