Update what is running queries to reflect new columns
This ended up being a bit of an overhaul, but its been tested across all
relevant AWS instances

1. Checks product versions the same way sp_Blitz does
2. Remove support for 2005 -- it won't compile with the new OFFSET data
type -- nothing will run there anyway
3. For 2008 and 2008R2, runs a stripped down version of a 'more
interesting' what is running query that for some reason wasn't used the
first time
4. for versions of 2012 and 2014 that don't have the new columns for:
mem grants etc, a case expression takes them out of the dynamic sql
5. for versions that do support it, they're included

Testing was good on this. Works across all versions, with confidence,
since our 2014 instance in AWS isn't patched to SP2, but my local is,
and the minor version of our 2016 instance is pre-RTM, and that worked.
2012 Worked including the columns, but there's no pre-SP2 handy to test
the without, but I think I got the comparison right on this one ;)