This adds the ability to sort by memory grant and average memory grant
This was a pain in the ass. Sorting by memory grant has similar
limitations to ignore query hashes/only query hashes. You can't examine
triggers or procs. They don't hit the right system view to use that
column. I also updated help and configuration table sections for this,
checks for it as a valid sort order.

This also adds a debugging section that prints out the troublesome
dynamic queries if you unquote it, and give you the data length of the
@sql variable. This really helped me figure out what was wrong here. It
can be taken out, but I'm leaving it just in case it sounds like a good
idea.

I also did some adding of N'' in front of sections that didn't have it.
SQL can behave oddly when you concatenate non unicode strings into
nvarchar(max) strings, where it converts them to nvarchar 4000 and can
truncate longer statements.

Anyway, yeah, this is working well and ready to test.