Add dm_os_waiting_tasks to wait stats
I wrote this query to ignore system sessions (id < 50), and added the
case logic we discussed to subtract and filter seconds depending on if
seconds = 0 or not. Query was written this way to cut down on filtering
out ignorable waits once from each session. I didn't avoid a CTE for any
reason. Whatever. I didn't modify any queries that reference the temp
table we're inserting into because of the ephemeral nature of the DMV.