Fooling SQL
This was interesting! The query that displays the wait stats information
as a separate box has a WHERE clause that looks for
(wd2.waiting_tasks_count - wd1.waiting_tasks_count) > 0, so I injected
some fake values to always get a value of 1. It does not work to SUM or
COUNT here, because rows disappear from this DMV, and it may have
nothing in it on the second go around.