Fixing an old(er) bug
There was a problem with the @Seconds = 0 gathering historical wait
stats query. Since we were assigning the system startup time to
@StartSampleTime at the beginning, when we inserted the value it was the
SAME for Pass 1 and Pass 2. This made the JOIN condition for the
@Seconds = 0 code invalid (wd2.SampleTime > wd1.SampleTime).

I changed it to wd2.SampleTime >= wd1.SampleTime, and fixed a divide by
zero error that was now returning a 0 for the DATEDIFF function.
Reserving one hug from @BrentOzar.