More work on the graceful failure
This was weird. On 2005, despite the RETURN; it would still print a wall
of red text afterwards for all the incompatible stuff in the stored
proc.

I tried TRY/CATCH and a GOTO to skip over the body of the stored proc,
but I couldn't get it working.

The other issue is that not everything would work going back to 2000, so
I applied copious duct tape and WD-40 to push the red text off the
screen instead. Really don't want to spend a ton of time on this one.