Update sp_BlitzFirst to allow @ShowSleepingSPIDs
Minor change to allow for the @ShowSleepingSPIDs paramater to be passed to sp_BlitzWho. Have changed the EXEC into dynamic SQL to allow for this. Same issue as with sp_BlitzWho with regards to data types. I've kept as a TINYINT but we're having to convert for the dynamic SQL to NVARCHAR.