Added parameterization and cleaned up some stuff
In this round:
Added dynamic sql params for database name and top in the all sort block
Added recompile hints where they were missing
Added catch if people try to get more than 10 plans per batch in all
sort mode
Changed how @Top is handled in the proc. It used to have a default
value, but now it's set based on sort order. Specifically, it's 10 if
we're not sorting by all, and it's 5 if we are, unless you set it
yourself.