Fun fact
sys.types has schema_id in it, but the only id in there is sys. So it
broke a bunch of stuff trying to join that to sys.schemas. Had to add in
a join to sys.objects. Also added debugging section for temp tables,
since we don't user balleriffic global temp tabled like BC.