Fixed Diff Restore, Added more Features
Updated the sp names for issue #851

Added @Debug variable.  Debug level 0 is the default state of the
script, which will execute the commands.  Debug level 1 will print just
the commands so you can run at your leisure and Debug level 2 will print
other stuff (level 2 is handy for adding in new variables and
troubleshooting).  This addresses issue #850

Fixed the Differential restore.  I discovered during testing that I had
failed to check the LSNs against the differential.  Fixed that portion
of the code and tested.  This is in regards to issue #814