Add some basic sanity checks for folder contents
If we can't find the path, or the folder is empty, etc., we should let
people know rather than just saying the command completed successfully.
It adds some redundancy to the polling process in sp_AllNightLog, but
this is good to have for people not running the two in conjunction, or
for file system changes in between folder discovery and a later restore
attempt.