Update for sp_foreachdb.sql
I have run into an issue where if I pass a @database_list longer than 18225 characters my query will not be run against all of the databases in the list even though the variable is an nvarchar(max).  Since SQL Server 2016 allows 32,767 databases I could see where this could happen to others.  I ran into the problem with 800 databases(not that I love having that many on one instance)  
There may be a better cleaner way to write this fix, but for me I ended up using Itzik Ben-Gan's auxillary table of numbers to create a temp table capping at 1/2 a million characters and referencing that table in the CTE versus using sysobjects the way it was before .  Perhaps you might want to increase the characters limit.