Changed output to honor session's regional setting
Changed the output to reflect current session's regional settings (regarding numeric format, decimal and grouping separator). 
The original CAST(AS MONEY) was extended to FORMAT(CAST(x AS MONEY), 'N'), making sure the number is formatted according to regional settings. 
The logic to strip zero-valued decimals to integers was corrected to support the now dynamic decimal separator, which in turn is determined by simply testing the output of a FORMAT statement. AFAIK there is no way to pull the current decimal separator from SQL Server's configuration. If there is, that method should be used here. 

Can be tested by changing the current session's language. For instance:
SET LANGUAGE Dutch;
Will set the session's language to Dutch (obviously) and therefore use a comma as decimal separator and period as grouping separator.