Update sp_ineachdb.sql
Changed 4 IIF statements to Case statements 
Changed 1 CONCAT statement to a ISNULL(C.Name,'') + ...

This also contains the 3 @NoSpace case sensitivity fix.

Brent and Team I left the previous 4 IIF statements and 1 CONCAT statement in the code but commented.  Feel free to remove them and/or use them to test.  I tested on SQL Server 2008 thru 2019.  As well as a case sensitive instance.  Feel free to reach out for any questions and/or concerns.