Working, probably
I spent a while on this trying to understand why wait stats + cpu time were coming back so much lower, and it looks like I don't need to do a first pass here. All I need to do is grab it on the second pass. That seems to accurately reflect waits + cpu.