Update sp_BlitzLock.sql
Rearrange the rollup info bar a little bit to include the version number instead of the version date, because it was unclear what was being shown there. Also adding in the start and end dates for the period of analysis, in case anyone thinks the results are incomplete, or since server startup.