Add some permissions checks to sp_Blitz
Closes #3292

This feels like a somewhat naïve set of checks. I don't know a ton about security.

It may clash when someone uses `EXECUTE AS` or signs the procedure for execution for lower-privileged users.

It's also incomplete at the moment, because I need to round up commands that touch system databases we may not have read permissions in.

I fully expect this to get rejected, but it got me error-free runs in a tightly locked down SQL Server environment.

If anyone has feedback, I'm happy to take it.