Fix date math and improve MI compatibility
This pull request looks a little funny, because when I re-synced my fork, the available plans query was all beat up for some reason, and I had to paste in the version from the current master branch.

Anyway, the two actual changes in this PR:
 * Closes #3385
 * Closes #3392

Since I can't repro #3385, I'm making the smallest change possible that fixes the issue with mocked-up test data. It will need some community testing to validate.

For #3392 I added a local variable to detect MI, and then set the TargetSessionType to ring_buffer if it's NULL, and we're looking at the system_health extended event. It might be a good idea to have this happen for any session on MI, but I'm going to keep it limited at first.